requirements.txt
Flask
flask-mongoengine
mongoengine
python-dotenv
Flask-JWT-Extended
Werkzeug
